chore: moved router events to core
