+++ packages/solid-crs-presentation/lib/app.events.ts
@@ -1,8 +1,8 @@
 import { Alert } from '@netwerk-digitaal-erfgoed/solid-crs-components';
-import { Collection } from '@netwerk-digitaal-erfgoed/solid-crs-core';
+import { Collection, RouterEvent } from '@netwerk-digitaal-erfgoed/solid-crs-core';
 import { DoneInvokeEvent, EventObject } from 'xstate';
 import { assign, choose, send } from 'xstate/lib/actions';
-import { AppContext, AppFeatureStates } from './app.machine';
+import { AppContext } from './app.machine';
 import { SearchUpdatedEvent } from './features/search/search.events';
 import { SelectedCollectionEvent } from 'features/collection/collection.events';
 import { SelectedObjectEvent } from 'features/object/object.events';
@@ -16,11 +16,7 @@ export enum AppEvents {
   ERROR = 'xstate.error',
   COLLECTIONS_LOADED = '[AppEvent: Collections loaded]',
   SET_PROFILE = '[AppEvent: Set Profile]',
-  NAVIGATE = '[AppEvent: Navigate]',
-  RESOLVED_ROUTE = '[AppEvent: Resolved Route]',
-  NAVIGATED = '[AppEvent: Navigated]',
   CLICKED_HOME = '[AppEvent: Clicked Home]',
-  POP_STATE = '[AppEvent: PopState]',
 }
 
 /**
@@ -69,36 +65,6 @@ export class SetProfileEvent implements EventObject {
 /**
  * An event which is dispatched when the user navigates
  */
-export class NavigateEvent implements EventObject {
-
-  public type: AppEvents.NAVIGATE = AppEvents.NAVIGATE;
-  constructor(public path?: string) { }
-
-}
-
-/**
- * An event which is dispatched when the user navigates
- */
-export class ResolvedRouteEvent implements EventObject {
-
-  public type: AppEvents.RESOLVED_ROUTE = AppEvents.RESOLVED_ROUTE;
-  constructor(public target: AppFeatureStates[keyof AppFeatureStates]) { }
-
-}
-
-/**
- * An event which is dispatched when the user has finished navigation
- */
-export class NavigatedEvent implements EventObject {
-
-  public type: AppEvents.NAVIGATED = AppEvents.NAVIGATED;
-  constructor(public path: string, public title?: string) { }
-
-}
-
-/**
- * An event which is dispatched when the user navigates
- */
 export class ClickedHomeEvent implements EventObject {
 
   public type: AppEvents.CLICKED_HOME = AppEvents.CLICKED_HOME;
@@ -106,19 +72,10 @@ export class ClickedHomeEvent implements EventObject {
 }
 
 /**
- * An event which is dispatched when this pod should be set up as an instition pod
- */
-export class PopStateEvent implements EventObject {
-
-  public type: AppEvents.POP_STATE = AppEvents.POP_STATE;
-  constructor(public event: PopStateEvent) { }
-
-}
-
-/**
  * Union type of app events.
  */
 export type AppEvent =
+  | RouterEvent
   | ErrorEvent
   | DismissAlertEvent
   | AddAlertEvent
@@ -126,11 +83,7 @@ export type AppEvent =
   | SelectedObjectEvent
   | SetProfileEvent
   | SearchUpdatedEvent
-  | NavigateEvent
-  | NavigatedEvent
-  | ResolvedRouteEvent
-  | ClickedHomeEvent
-  | PopStateEvent;
+  | ClickedHomeEvent;
 
 /**
  * Actions for the alerts component.
